Refactor computation of signed serial
